Assumes ThumbnailLoadJob::appendItem() is only called on image files (coherent with previous change).
svn path=/trunk/kdeextragear-1/gwenview/; revision=339074

+++ src/gvfilethumbnailview.cpp
@@ -302,6 +302,9 @@ void GVFileThumbnailView::slotUpdateEnded() {
 
 
 void GVFileThumbnailView::updateThumbnail(const KFileItem* fileItem) {
+	if (fileItem->isDir() || GVArchive::fileItemIsArchive(fileItem)) {
+		return;
+	}
 
 	ThumbnailLoadJob::deleteImageThumbnail(fileItem->url());
 	if (d->mThumbnailLoadJob.isNull()) {
